[REFACTOR #42]: DTO 보완 again (#45)
* refactor: field type, name

* Update GradeResponse.java

+++ src/main/java/com/payper/benefit/dto/DiscountResponse.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @Builder
 public class DiscountResponse {
 	private String type;
-	private int amount;
-	private int limitCount;
-	private int limitAmount;
+	private Long amount;
+	private Long limitCount;
+	private Long limitAmount;
 }
